requirements.txt
awesome-slugify==1.6.5
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
djangorestframework==3.10.3
idna==2.8
oauthlib==3.1.0
Pillow==6.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
Unidecode==0.4.21
urllib3==1.25.6
